// api/model/roleCreate.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * AI Shop Helper
 * OpenAPI spec version: 0.1.0
 */

/**
 * Schema for creating a new role.
 */
export interface RoleCreate {
  /**
   * @minLength 1
   * @maxLength 255
   */
  name: string;
  description?: string | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  access_level: number;
}