// api/model/rolePublic.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * AI Shop Helper
 * OpenAPI spec version: 0.1.0
 */

/**
 * Schema for returning role data to the client.
 */
export interface RolePublic {
  id: string;
  name: string;
  description: string | null;
  access_level: number;
  created_at: string;
  updated_at: string;
  created_by: string;
  updated_by: string;
}